+ Vue.js

+++ src/main/java/com/FreeDoska/doska/AnnouncementModel.java
@@ -1,7 +1,9 @@
 package com.FreeDoska.doska;
+
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.ToString;
+
 import javax.persistence.*;
 import java.time.LocalDateTime;
 
@@ -13,7 +15,7 @@ import java.time.LocalDateTime;
 public class AnnouncementModel {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer id;
+    private Long id;
     private String nameAnnouncement;
     private String textAnnouncement;
     private String category;
@@ -25,12 +27,11 @@ public class AnnouncementModel {
     private LocalDateTime creationDate;
 
 
-
-    public long getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
