+++ src/main/java/com/FreeDoska/doska/repo/AnnouncementRepo.java
@@ -1,8 +1,10 @@
 package com.FreeDoska.doska.repo;
-import com.FreeDoska.doska.AnnouncementModel;
 
-import org.springframework.data.repository.CrudRepository;
+import com.FreeDoska.doska.AnnouncementModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AnnouncementRepo extends JpaRepository<AnnouncementModel, Integer> {
+import java.util.List;
+
+public interface AnnouncementRepo extends JpaRepository<AnnouncementModel, Long> {
+//List<AnnouncementModel> findByTag(String tag);
 }
